requirements.txt
# Core ADK and AI dependencies
google-adk==1.22.1
google-genai==1.24.0
google-generativeai==0.8.5

# Memory system
mem0ai==0.1.114  # For LTM vector/graph memory
chromadb==0.6.3  # Required for Mem0 local storage

# Audio processing
openai-whisper==20231117  # For audio transcription
soundfile==0.12.1  # For audio file handling

# Telegram bot
python-telegram-bot==22.2

# Environment and configuration
python-dotenv==1.1.1

# Web framework (for ADK web UI)
fastapi==0.115.14
uvicorn==0.35.0

# Database
SQLAlchemy==2.0.41
aiosqlite==0.22.1

# HTTP and networking
httpx==0.28.1
requests==2.32.4

# Utilities
python-multipart==0.0.20
pytest==8.2.2

# Web Search
googlesearch-python==1.2.3

# Vector embeddings (for future use)
sentence-transformers==2.2.2  # For semantic embeddings
numpy==1.24.3  # Required for embeddings

# Observability
freeplay-python-adk==0.2.2  # For ADK observability
